+++ packages/cli/package.json
@@ -34,7 +34,6 @@
     "@pikku/better-auth": "^0.12.36",
     "@pikku/bun-server": "^0.12.11",
     "@pikku/code-edit": "^0.12.1",
-    "@pikku/db-migrator": "^0.12.1",
     "@pikku/deploy": "^0.12.5",
     "@pikku/deploy-cloudflare": "^0.12.13",
     "@pikku/fetch": "^0.12.10",
@@ -46,6 +45,7 @@
     "@pikku/openapi-parser": "^0.12.20",
     "@pikku/schedule": "^0.12.7",
     "@pikku/skills": "^0.12.23",
+    "@pikku/sql-migrator": "^0.12.1",
     "@types/cookie": "^1.0.0",
     "@types/json-schema": "^7.0.15",
     "ai": "^7.0.66",
