+++ packages/cli/src/functions/db/postgres/postgres-introspector.ts
@@ -5,7 +5,7 @@ import type {
   ForeignKeyInfo,
   EnumInfo,
 } from '../db-introspector.js'
-import { MIGRATION_TRACKING_TABLE } from '@pikku/db-migrator'
+import { MIGRATION_TRACKING_TABLE } from '@pikku/sql-migrator'
 import { parseCheckEnumValues } from '../check-enums.js'
 
 interface PgColumnRow {
