+++ packages/deploy/deploy-standalone/package.json
@@ -18,7 +18,7 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@pikku/db-migrator": "^0.12.1",
-    "@pikku/deploy": "^0.12.5"
+    "@pikku/deploy": "^0.12.5",
+    "@pikku/sql-migrator": "^0.12.1"
   }
 }
