+++ packages/sql-migrator/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@pikku/db-migrator",
+  "name": "@pikku/sql-migrator",
   "version": "0.12.1",
   "description": "The SQL migration applier shared by the Pikku CLI and the standalone runtime",
   "author": "yasser.fadl@gmail.com",
