+++ packages/sql-migrator/run-tests.sh
