+++ packages/sql-migrator/src/postgres/postgres-migrator.ts
@@ -1,5 +1,5 @@
-import type { MigrationExecutor, AppliedMigration } from '../db-migrator.js'
-import { MIGRATION_TRACKING_TABLE as TRACKING_TABLE } from '../db-migrator.js'
+import type { MigrationExecutor, AppliedMigration } from '../sql-migrator.js'
+import { MIGRATION_TRACKING_TABLE as TRACKING_TABLE } from '../sql-migrator.js'
 
 export interface PostgresMigrationClient {
   query<T = unknown>(sql: string, params?: unknown[]): Promise<{ rows: T[] }>
