+++ packages/sql-migrator/tsconfig.json
